:wrench: fix: Fixed sign out method
Fixed sign out method by fixing revokeToken and isTokenRevoked at RedisService
The method isTokenRevoked was using the wrong method to find out if token was in redis or not. Fixed by change ISMEMBERS to EXISTS
The method revokeToken was needing a guarantee to ttl be always positve also revokeToken was black listing tokens at a different key.